Rename formatters to transformers
