--- src/utils/clients/discogs/methods/index.ts
@@ -1,1 +1,0 @@
-export * from './album-search';
